Creates a GUI to ob used as an import
See: SYNOPSIS 

Used for the administration of Exchange. Most commonly used in Distribution List management but the usefulness extends to any cmdlet supporting Email or principal name formats for Identity. 

Once the function is loaded in the ISE, you can use it with PIPE commands. Like

Parse-emails | Get-mailbox #Gets all mailboxes
Parse-emails | Remove-distributiongroupmember -Idenity "MYDL" 

Normally, a user list would have to be saved to a CSV file then imported to a variable. Having a GUI saves several repetitive steps.

NOTE: Winforms work better within MTA runspaces but this script manages the runsapce to deal with it. 

Works with PS5, not tested with older versions