Create Example of a Dynamic Parameter
A Fun and practical example of how to define a Dynamic Parameter in a function. This can be used in scriptwriting to provide choices based on a result. For example, I have a Dynamic Parameter in a custom "Move-ActiveMailboxDatabase" which shows the existing Databases then presents the servers to move to by Activation Preference.